ai-gateway/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
litellm==1.76.1
python-dotenv==1.0.1
